requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
idna==2.6
praw==5.3.0
prawcore==0.13.0
python-dotenv==0.7.1
requests==2.18.4
update-checker==0.16
urllib3==1.22
